Add a menu to choose the game to print the variance for, and only use this. Also added a horizontal scrollbar on the viewing area and anchors the entries to the left.
