+++ varianceApp.py
@@ -3,6 +3,8 @@
 import tkinter as tk
 import threading
 from PracticeComponents import VarianceColumn
+from DataClasses import AllSplitNames
+import varianceCalculator as varcalc
 
 class App(threading.Thread):
     def __init__(self):
@@ -23,6 +25,23 @@ class App(threading.Thread):
         self.base.pack(side="left")
         self.sort.pack(side="right")
 
+        self.splits = AllSplitNames.Splits()
+        menubar = tk.Menu(self.root, tearoff=False)
+        fileMenu = tk.Menu(self.root, tearoff=False)
+
+        menubar.add_cascade(label="Run Reference", menu=fileMenu)
+        lambdas = [[lambda game=game, category=category: self.setRun(game,category) for category in self.splits.getCategories(game)] for game in self.splits.getGames()]
+        i = 0
+        for game in self.splits.getGames():
+            recentMenu = tk.Menu(self.root, tearoff=False)
+            fileMenu.add_cascade(label=game,menu=recentMenu)
+            j = 0
+            for category in self.splits.getCategories(game):
+                recentMenu.add_command(label=category,command=lambdas[i][j])
+                j = j + 1
+            i = i + 1
+        self.root.configure(bg="black",menu=menubar)
+
     ##########################################################
     ## Show the window, and call the first update after one
     ## frame.
@@ -33,3 +52,7 @@ class App(threading.Thread):
     def showVariances(self,variance,sort):
         self.base.update(variance)
         self.sort.update(sort)
+
+    def setRun(self,game,category):
+        variances = varcalc.computeVariances(game,category,self.splits)
+        self.showVariances(variances["order"],variances["sorted"])
